refactor: modularize background and offscreen (#35)
