+++ src/entries/Background/mutex.ts
@@ -1,0 +1,5 @@
+import { Mutex } from 'async-mutex';
+
+const mutex = new Mutex();
+
+export default mutex;
