--- wasm/prover/check.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-cargo check --target "wasm32-unknown-unknown"
